+++ NodeJs-Archive-Tool/lib/lock/archiveUnsetLock.js
@@ -2,6 +2,8 @@ const fs = require('fs')
 const archiveGetLockName = require('./archiveGetLockName.js')
 
 module.exports = function (archiveFormat) {
-  fs.unlinkSync(archiveGetLockName(archiveFormat))
+  if (fs.existsSync(archiveGetLockName(archiveFormat))) {
+    fs.unlinkSync(archiveGetLockName(archiveFormat))
+  }
   return true
 }
